Project 1/requirements.txt
agate==1.7.1
aiobotocore==2.12.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
boto3==1.34.51
botocore==1.34.51
build==1.1.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==14.0
crashtest==0.4.1
croniter==2.0.2
cryptography==42.0.5
dagster==1.6.5
dagster-dbt==0.22.5
dagster-docker==0.22.5
dagster-graphql==1.6.5
dagster-pipes==1.6.5
dagster-postgres==0.22.5
dagster-webserver==1.6.5
dbt-bigquery==1.7.6
dbt-core==1.7.9
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.4
distlib==0.3.8
docker==7.0.0
docker-image-py==0.1.12
docstring-parser==0.15
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud==0.34.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.9.3
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
installer==0.7.0
isodate==0.6.1
jaraco.classes==3.3.1
jeepney==0.8.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.1
leather==0.4.0
Logbook==1.5.3
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mashumaro==3.12
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.2.0
msal==1.27.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
orjson==3.9.15
packaging==24.0
pandas==2.2.1
parsedatetime==2.6
pathspec==0.11.2
pendulum==3.0.0
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3fs==2024.2.0
s3transfer==0.10.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
starlette==0.37.2
structlog==24.1.0
tabulate==0.9.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.4
toposort==1.10
tqdm==4.66.2
trove-classifiers==2024.3.3
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
universal_pathlib==0.2.2
urllib3==1.26.18
uvicorn==0.28.0
uvloop==0.19.0
virtualenv==20.25.1
watchdog==4.0.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
